frontend/requirements.txt
streamlit
plotly
google-cloud-aiplatform
python-dotenv
setuptools